Fixed GWS issue
The major "Gender and Women's Studies" was not able to be joined because of the apostrophe. Server-side, this role was renamed to remove the apostrophe. The corresponding entries on message.txt and majors.py have been changed to reflect this update. Also implemented scrubbing of apostrophe's on user input for major, which means the command "$major gender and women's studies" will still allow a user to join the correct major.