Fixed $major Exploit
The Classes cog now builds a list of majors from message.txt. This is used by the $major command to validate the user's input for their major. This means any arbitrary role can be added, and any non-major role can be renamed without worry of that role becoming joinable through the $major command.